Moved seniority level from description to position :facepalm:
