+++ php/hireinsocial/db/orm/mapping/xml/Offer.Position.orm.xml
@@ -3,6 +3,7 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                           https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
     <embeddable name="HireInSocial\Offers\Application\Offer\Position">
+        <field name="seniorityLevel" type="smallint"/>
         <field name="name"/>
         <field name="description" length="1024"/>
     </embeddable>
