+++ php/hireinsocial/resources/templates/en-US/ui/offers/offer/new.html.twig
@@ -51,6 +51,7 @@
             {{ form_row(form.company.description, {label: "Description", attr:{rows: 5, placeholder: "between 10 and 512 characters..."}}) }}
             <hr />
             <h5>Position</h5>
+            {{ form_row(form.position.seniority_level, {label: "Seniority Level"}) }}
             {{ form_row(form.position.name, {label: "Name", attr:{placeholder: "PHP Developer"}}) }}
             {{ form_row(form.position.description, {label: "Description", attr:{rows: 5, placeholder: "between 50 and 1024 characters..."}}) }}
             <hr />
@@ -90,7 +91,6 @@
             <hr/>
 
             <h5>Description</h5>
-            {{ form_row(form.description.seniority_level, {label: "Seniority Level"}) }}
             {{ form_row(form.description.requirements, {label: "Requirements", attr:{rows: 5, placeholder: "between 100 and 1024 characters..."}}) }}
             {{ form_row(form.description.benefits, {label: "Benefits", attr:{rows: 5, placeholder: "between 100 and 1024 characters..."}}) }}
             <hr />
