+++ php/hireinsocial/src/App/Offers/Form/Type/Offer/DescriptionType.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace App\Offers\Form\Type\Offer;
 
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Validator\Constraints\Length;
@@ -24,18 +23,6 @@ final class DescriptionType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options) : void
     {
         $builder
-            ->add('seniority_level', ChoiceType::class, [
-                'required' => true,
-                'expanded' => true,
-                'multiple' => false,
-                'choices' => [
-                    'Intern'  => 0,
-                    'Junior'  => 1,
-                    'Mid'  => 2,
-                    'Senior'  => 3,
-                    'Expert'  => 4,
-                ],
-            ])
             ->add('requirements', TextareaType::class, [
                 'constraints' => [
                     new Length(['min' => 100, 'max' => 1024]),
