+++ php/hireinsocial/src/HireInSocial/Offers/Application/Offer/Position.php
@@ -14,9 +14,15 @@ declare(strict_types=1);
 namespace HireInSocial\Offers\Application\Offer;
 
 use HireInSocial\Offers\Application\Assertion;
+use HireInSocial\Offers\Application\Offer\Position\SeniorityLevels;
 
 final class Position
 {
+    /**
+     * @var int
+     */
+    private $seniorityLevel;
+
     /**
      * @var string
      */
@@ -27,13 +33,23 @@ final class Position
      */
     private $description;
 
-    public function __construct(string $name, string $description)
+    public function __construct(int $seniorityLevel, string $name, string $description)
     {
+        Assertion::inArray($seniorityLevel, SeniorityLevels::all());
         Assertion::betweenLength($name, 3, 255);
         Assertion::betweenLength($description, 50, 1024);
 
+        $this->seniorityLevel = $seniorityLevel;
         $this->name = $name;
         $this->description = $description;
+    }
+
+    /**
+     * @return int
+     */
+    public function seniorityLevel() : int
+    {
+        return $this->seniorityLevel;
     }
 
     public function name() : string
