+++ php/hireinsocial/src/HireInSocial/Offers/Application/Query/Offer/Model/Offer/Description.php
@@ -13,15 +13,8 @@ declare(strict_types=1);
 
 namespace HireInSocial\Offers\Application\Query\Offer\Model\Offer;
 
-use HireInSocial\Offers\Application\Exception\Exception;
-
 final class Description
 {
-    /**
-     * @var int
-     */
-    private $seniorityLevel;
-
     /**
      * @var string
      */
@@ -32,29 +25,10 @@ final class Description
      */
     private $benefits;
 
-    public function __construct(int $seniorityLevel, string $requirements, string $benefits)
+    public function __construct(string $requirements, string $benefits)
     {
-        $this->seniorityLevel = $seniorityLevel;
         $this->requirements = $requirements;
         $this->benefits = $benefits;
-    }
-
-    public function seniorityLevel() : string
-    {
-        switch ($this->seniorityLevel) {
-            case 0:
-                return 'Intern';
-            case 1:
-                return 'Junior';
-            case 2:
-                return 'Mid';
-            case 3:
-                return 'Senior';
-            case 4:
-                return 'Expert';
-            default:
-                throw new Exception("Unknown seniority level");
-        }
     }
 
     public function requirements() : string
