+++ php/hireinsocial/tests/App/Tests/Offers/Functional/Web/OfferTest.php
@@ -67,6 +67,7 @@ final class OfferTest extends WebTestCase
             'offer[company][name]' => 'Company name',
             'offer[company][url]' => 'http://company.com',
             'offer[company][description]' => $faker->text(512),
+            'offer[position][seniority_level]' => \random_int(0, 4),
             'offer[position][name]' => 'Software Developer',
             'offer[position][description]' => $faker->text(1024),
             'offer[salary][min]' => 1000,
@@ -78,7 +79,6 @@ final class OfferTest extends WebTestCase
             'offer[location][name]' => 'Cracow',
             'offer[location][lat]' => '50.06212',
             'offer[location][lng]' => '19.9353153',
-            'offer[description][seniority_level]' => \random_int(0, 4),
             'offer[description][requirements]' => $faker->text(1024),
             'offer[description][benefits]' => $faker->text(1024),
             'offer[contact][email]' => $faker->email,
