+++ php/hireinsocial/tests/HireInSocial/Tests/Offers/Application/MotherObject/Offer/OfferMother.php
@@ -36,12 +36,11 @@ final class OfferMother
             SpecializationMother::random(),
             UserMother::random(),
             new Offer\Company($companyName, $faker->url, $faker->text(512)),
-            new Offer\Position($positionName, $faker->text(1024)),
+            new Offer\Position(\random_int(Offer\Position\SeniorityLevels::INTERN, Offer\Position\SeniorityLevels::EXPERT), $positionName, $faker->text(1024)),
             Offer\Location::onlyRemote(),
             new Offer\Salary($faker->numberBetween(1000, 5000), $faker->numberBetween(5000, 20000), 'PLN', $faker->boolean, Offer\Salary\Period::perMonth()),
             new Offer\Contract('B2B'),
             new Offer\Description(
-                \random_int(Offer\Description\SeniorityLevels::INTERN, Offer\Description\SeniorityLevels::EXPERT),
                 $faker->text(1024),
                 $faker->text(1024)
             ),
